user-service/requirements.txt
Flask>=2.0.0
SQLAlchemy>=1.4.0
PyJWT>=2.0.0
bcrypt>=3.2.0
python-json-logger>=2.0.0